# src/NPaperless.Services/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/saved_views/" : {
      "get" : {
        "operationId" : "GetSavedViews",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "page_size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":6,\"next\":null,\"previous\":null,\"all\":[8,17,7,4,11,15],\"results\":[{\"id\":8,\"name\":\"Correspondent 2\",\"show_on_dashboard\":false,\"show_in_sidebar\":false,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[{\"rule_type\":3,\"value\":\"2\"}],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true},{\"id\":17,\"name\":\"In the Last Month\",\"show_on_dashboard\":false,\"show_in_sidebar\":false,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[{\"rule_type\":20,\"value\":\"created:[-1 month to now]\"}],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true},{\"id\":7,\"name\":\"Inbox\",\"show_on_dashboard\":true,\"show_in_sidebar\":true,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[{\"rule_type\":6,\"value\":\"9\"}],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true},{\"id\":4,\"name\":\"Recently Added\",\"show_on_dashboard\":true,\"show_in_sidebar\":true,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true},{\"id\":11,\"name\":\"Tag: Another Sample Tag\",\"show_on_dashboard\":false,\"show_in_sidebar\":true,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[{\"rule_type\":6,\"value\":\"4\"}],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true},{\"id\":15,\"name\":\"View ASN not empty\",\"show_on_dashboard\":false,\"show_in_sidebar\":false,\"sort_field\":\"created\",\"sort_reverse\":true,\"filter_rules\":[{\"rule_type\":18,\"value\":\"false\"}],\"owner\":{\"id\":2,\"password\":\"pbkdf2_sha256$390000$ElHGEQzOIS3Gv2wu8YDbSt$bWASwmUfHtFMZ6+rG+4rrrtejjC9RiwXm6xMi0bxMZo=\",\"last_login\":\"2023-04-26T09:08:14Z\",\"is_superuser\":false,\"username\":\"testuser\",\"first_name\":\"\",\"last_name\":\"\",\"email\":\"test@user.com\",\"is_staff\":true,\"is_active\":true,\"date_joined\":\"2021-01-03T07:37:04Z\",\"groups\":[],\"user_permissions\":[57,58,59,60,5,6,7,8,1,2,3,4,9,10,11,12,61,62,63,64,65,66,67,68,13,14,15,16,117,118,119,120,121,122,123,124,113,114,115,116,77,78,79,80,85,86,87,88,69,70,71,72,81,82,83,84,73,74,75,76,33,34,35,36,21,22,23,24,37,38,39,40,89,90,91,92,29,30,31,32,109,137,110,138,111,139,112,140,105,106,107,108,41,42,43,44,45,46,47,48,93,94,95,96,25,26,27,28,101,102,103,104,97,98,99,100,125,126,127,128,129,130,131,132,49,50,51,52,53,54,55,56,133,134,135,136,17,18,19,20]},\"user_can_change\":true}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetSavedViews_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      },
      "post" : {
        "operationId" : "CreateSavedViews",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"test\",\"show_on_dashboard\":true,\"show_in_sidebar\":true,\"filter_rules\":[{\"rule_type\":19,\"value\":\"test\"},{\"rule_type\":28,\"value\":\"1\"},{\"rule_type\":30,\"value\":\"1\"}],\"sort_reverse\":true,\"sort_field\":\"created\"}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/CreateSavedViews_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      }
    },
    "/api/storage_paths/" : {
      "get" : {
        "operationId" : "GetStoragePaths",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "full_perms",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":1,\"next\":null,\"previous\":null,\"all\":[1],\"results\":[{\"id\":1,\"slug\":\"testdir\",\"name\":\"testdir\",\"path\":\"testdirp\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"document_count\":1,\"owner\":3,\"permissions\":{\"view\":{\"users\":[3],\"groups\":[]},\"change\":{\"users\":[3],\"groups\":[]}}}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetStoragePaths_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      },
      "post" : {
        "operationId" : "CreateStoragePath",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"testdeldir\",\"path\":\"testdeldir\",\"matching_algorithm\":6,\"match\":\"\",\"is_insensitive\":true,\"owner\":3}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/CreateStoragePath_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":2,\"slug\":\"testdeldir\",\"name\":\"testdeldir\",\"path\":\"testdeldir\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"owner\":3,\"user_can_change\":true}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/CreateStoragePath_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      }
    },
    "/api/storage_paths/{id}/" : {
      "delete" : {
        "operationId" : "DeleteStoragePath",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      },
      "put" : {
        "operationId" : "UpdateStoragePath",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"id\":1,\"slug\":\"testdir\",\"name\":\"testdir2\",\"path\":\"testdirp\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"document_count\":1,\"owner\":3,\"permissions\":{\"view\":{\"users\":[3],\"groups\":[]},\"change\":{\"users\":[3],\"groups\":[]}},\"set_permissions\":{\"view\":{\"users\":[3],\"groups\":[1]},\"change\":{\"users\":[3],\"groups\":[1]}},\"permissions_form\":{\"owner\":3,\"set_permissions\":{\"view\":{\"users\":[3],\"groups\":[]},\"change\":{\"users\":[3],\"groups\":[]}}}}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/UpdateStoragePath_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":1,\"slug\":\"testdir2\",\"name\":\"testdir2\",\"path\":\"testdirp\",\"match\":\"\",\"matching_algorithm\":6,\"is_insensitive\":true,\"document_count\":1,\"owner\":3,\"user_can_change\":true}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/UpdateStoragePath_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GetSavedViews_200_response_results_inner_filter_rules_inner" : {
        "example" : {
          "rule_type" : 2,
          "value" : "value"
        },
        "properties" : {
          "rule_type" : {
            "type" : "integer"
          },
          "value" : {
            "type" : "string"
          }
        },
        "required" : [ "rule_type", "value" ],
        "type" : "object"
      },
...
      "GetSavedViews_200_response" : {
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : {
              "is_superuser" : true,
              "is_active" : true,
              "user_permissions" : [ 9, 9 ],
              "is_staff" : true,
              "last_login" : "last_login",
              "last_name" : "last_name",
              "groups" : [ "", "" ],
              "password" : "password",
              "id" : 7,
              "date_joined" : "date_joined",
              "first_name" : "first_name",
              "email" : "email",
              "username" : "username"
            },
            "user_can_change" : true,
            "sort_field" : "sort_field",
            "show_on_dashboard" : true,
            "name" : "name",
            "show_in_sidebar" : true,
            "filter_rules" : [ {
              "rule_type" : 2,
              "value" : "value"
            }, {
              "rule_type" : 2,
              "value" : "value"
            } ],
            "sort_reverse" : true,
            "id" : 5
          }, {
            "owner" : {
              "is_superuser" : true,
              "is_active" : true,
              "user_permissions" : [ 9, 9 ],
              "is_staff" : true,
              "last_login" : "last_login",
              "last_name" : "last_name",
              "groups" : [ "", "" ],
              "password" : "password",
              "id" : 7,
              "date_joined" : "date_joined",
              "first_name" : "first_name",
              "email" : "email",
              "username" : "username"
            },
            "user_can_change" : true,
            "sort_field" : "sort_field",
            "show_on_dashboard" : true,
            "name" : "name",
            "show_in_sidebar" : true,
            "filter_rules" : [ {
              "rule_type" : 2,
              "value" : "value"
            }, {
              "rule_type" : 2,
              "value" : "value"
            } ],
            "sort_reverse" : true,
            "id" : 5
          } ]
        },
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "items" : {
              "type" : "integer"
            },
            "type" : "array"
          },
          "results" : {
            "items" : {
              "$ref" : "#/components/schemas/GetSavedViews_200_response_results_inner"
            },
            "type" : "array"
          }
        },
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object"
      },
      "CreateSavedViews_request" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "show_on_dashboard" : {
            "type" : "boolean"
          },
          "show_in_sidebar" : {
            "type" : "boolean"
          },
          "filter_rules" : {
            "items" : {
              "$ref" : "#/components/schemas/GetSavedViews_200_response_results_inner_filter_rules_inner"
            },
            "type" : "array"
          },
          "sort_reverse" : {
            "type" : "boolean"
          },
          "sort_field" : {
            "type" : "string"
          }
        },
        "required" : [ "filter_rules", "name", "show_in_sidebar", "show_on_dashboard", "sort_field", "sort_reverse" ],
        "type" : "object"
      },
...
      "GetStoragePaths_200_response" : {
        "example" : {
          "next" : true,
          "all" : [ 6, 6 ],
          "previous" : true,
          "count" : 0,
          "results" : [ {
            "owner" : 2,
            "path" : "path",
            "matching_algorithm" : 5,
            "document_count" : 5,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ 7, 7 ]
              },
              "change" : {
                "groups" : [ "", "" ],
                "users" : [ 7, 7 ]
              }
            },
            "name" : "name",
            "match" : "match",
            "id" : 1,
            "slug" : "slug"
          }, {
            "owner" : 2,
            "path" : "path",
            "matching_algorithm" : 5,
            "document_count" : 5,
            "is_insensitive" : true,
            "permissions" : {
              "view" : {
                "groups" : [ "", "" ],
                "users" : [ 7, 7 ]
              },
              "change" : {
                "groups" : [ "", "" ],
                "users" : [ 7, 7 ]
              }
            },
            "name" : "name",
            "match" : "match",
            "id" : 1,
            "slug" : "slug"
          } ]
        },
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "boolean"
          },
          "previous" : {
            "type" : "boolean"
          },
          "all" : {
            "items" : {
              "type" : "integer"
            },
            "type" : "array"
          },
          "results" : {
            "items" : {
              "$ref" : "#/components/schemas/GetStoragePaths_200_response_results_inner"
            },
            "type" : "array"
          }
        },
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object"
      },
      "CreateStoragePath_request" : {
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "matching_algorithm" : {
            "type" : "integer"
          },
          "match" : {
            "type" : "string"
          },
          "is_insensitive" : {
            "type" : "boolean"
          },
          "owner" : {
            "type" : "integer"
          }
        },
        "required" : [ "is_insensitive", "match", "matching_algorithm", "name", "owner", "path" ],
        "type" : "object"
      },
...
    }
  }
}